Rust: Refactor Tranform By Direction
